+++ CHANGES.rst
@@ -11,6 +11,7 @@ LATEST
 - Fixed internal state of RadioButton values to be consistent with selection.
 - Fixed handling of zero width modifiers.
 - Fixed image conversion to use modern PIL API and sort off-by-one height error.
+- Fixed parser bug returning list instead of colour tuple.
 
 1.14.0
 ------